Split US and GB Components into different directories. (#41)
