+++ src/CovidApp.tsx
@@ -8,11 +8,11 @@ import {Dimensions, StatusBar} from "react-native";
 import { colors } from "../theme/colors";
 import { RegisterScreen } from "./features/register/RegisterScreen";
 import AboutYouScreen from "./features/patient/AboutYouScreen";
-import { WelcomeScreen } from "./features/register/WelcomeScreen";
+import { WelcomeScreen } from "./features/register/gb/WelcomeScreen";
 import { SplashScreen } from "./features/SplashScreen";
 import YourHealthScreen from "./features/assessment/YourHealthScreen";
 import HowYouFeelScreen from "./features/assessment/HowYouFeelScreen";
-import { PrivacyPolicyUKScreen } from "./features/register/PrivacyPolicyUKScreen";
+import { PrivacyPolicyUKScreen } from "./features/register/gb/PrivacyPolicyUKScreen";
 import { TermsScreen } from "./features/register/TermsScreen";
 import { OptionalInfoScreen } from "./features/register/OptionalInfoScreen";
 import CovidTestScreen from './features/assessment/CovidTestScreen';
@@ -22,16 +22,16 @@ import TreatmentSelectionScreen from './features/assessment/TreatmentSelectionSc
 import TreatmentOtherScreen from './features/assessment/TreatmentOtherScreen';
 import { ScreenParamList } from './features/ScreenParamList';
 import ThankYouScreen from './features/ThankYouScreen';
-import { WelcomeRepeatScreen } from "./features/register/WelcomeRepeatScreen";
-import { Welcome1USScreen } from "./features/register/Welcome1USScreen";
-import { Welcome2USScreen } from "./features/register/Welcome2USScreen";
-import { WelcomeRepeatUSScreen } from "./features/register/WelcomeRepeatUSScreen";
-import TermsOfUseScreen from "./features/register/TermsOfUseScreen";
-import { PrivacyPolicyUSScreen } from "./features/register/PrivacyPolicyUSScreen";
+import { WelcomeRepeatScreen } from "./features/register/gb/WelcomeRepeatScreen";
+import { Welcome1USScreen } from "./features/register/us/Welcome1USScreen";
+import { Welcome2USScreen } from "./features/register/us/Welcome2USScreen";
+import { WelcomeRepeatUSScreen } from "./features/register/us/WelcomeRepeatUSScreen";
+import TermsOfUseScreen from "./features/register/us/TermsOfUseScreen";
+import { PrivacyPolicyUSScreen } from "./features/register/us/PrivacyPolicyUSScreen";
 import {CountrySelectScreen} from "./features/CountrySelectScreen";
 import YourWorkScreen from './features/patient/YourWorkScreen';
-import {NursesConsentUSScreen} from "./features/register/NursesConsentUS";
-import BeforeWeStartUS from "./features/register/BeforeWeStartUS";
+import {NursesConsentUSScreen} from "./features/register/us/NursesConsentUS";
+import BeforeWeStartUS from "./features/register/us/BeforeWeStartUS";
 import {ResetPasswordScreen} from "./features/password-reset/ResetPassword";
 import {ResetPasswordConfirmScreen} from "./features/password-reset/ResetPassordConfirm";
 import HealthWorkerExposureScreen from "./features/assessment/HealthWorkerExposureScreen";
