+++ src/features/register/us/BeforeWeStartUS.tsx
@@ -1,13 +1,13 @@
 import React, {Component} from "react";
 import {StyleSheet} from "react-native";
-import Screen, {FieldWrapper, Header, screenWidth} from "../../components/Screen";
-import {HeaderText} from "../../components/Text";
+import Screen, {FieldWrapper, Header, screenWidth} from "../../../components/Screen";
+import {HeaderText} from "../../../components/Text";
 import {Form, Text} from "native-base";
 
-import {colors, fontStyles} from "../../../theme"
+import {colors, fontStyles} from "../../../../theme"
 import {StackNavigationProp} from "@react-navigation/stack";
-import {ScreenParamList} from "../ScreenParamList";
-import {BigButton} from "../../components/Button";
+import {ScreenParamList} from "../../ScreenParamList";
+import {BigButton} from "../../../components/Button";
 
 
 type HowYouFeelProps = {
