+++ src/features/register/us/TermsOfUseScreen.tsx
@@ -1,9 +1,9 @@
 import {StackNavigationProp} from "@react-navigation/stack";
 import React, {Component} from "react";
 import {Linking, ScrollView, StyleSheet, View} from "react-native";
-import {colors} from "../../../theme";
-import {BrandedButton, ClickableText, RegularBoldText, RegularText} from "../../components/Text";
-import {ScreenParamList} from "../ScreenParamList";
+import {colors} from "../../../../theme";
+import {BrandedButton, ClickableText, RegularBoldText, RegularText} from "../../../components/Text";
+import {ScreenParamList} from "../../ScreenParamList";
 
 type PropsType = {
     navigation: StackNavigationProp<ScreenParamList, 'Terms'>
